refactor:(server): add xml comments and format
