--- src/FinnHub.MCP.Server.Application/Common/ResultErrorType.cs
@@ -1,21 +1,0 @@
-﻿// ---------------------------------------------------------------------------------------------------------------------
-//  <copyright>
-//    This file is part of FinnHub MCP Server and is licensed under the MIT License.
-//    See the LICENSE file in the project root for full license information.
-//  </copyright>
-//  <summary>
-//    // TODO Add summary
-//  </summary>
-// ---------------------------------------------------------------------------------------------------------------------
-
-namespace FinnHub.MCP.Server.Application.Common;
-
-public enum ResultErrorType
-{
-    NotFound,
-    Unknown,
-    InvalidQuery,
-    ServiceUnavailable,
-    Timeout,
-    InvalidResponse
-}
