+++ src/FinnHub.MCP.Server.Application/Search/Features/SearchSymbol/StockSymbol.cs
@@ -3,9 +3,6 @@
 //    This file is part of FinnHub MCP Server and is licensed under the MIT License.
 //    See the LICENSE file in the project root for full license information.
 //  </copyright>
-//  <summary>
-//    // TODO Add summary
-//  </summary>
 // ---------------------------------------------------------------------------------------------------------------------
 
 using System.ComponentModel;
